Arranged components into folders
