+++ react-ui/src/App.js
@@ -9,22 +9,22 @@ import { initializeEquipment } from './reducers/equipmentReducer'
 import { initializeRentals } from './reducers/rentalReducer'
 import NavBar from './components/structure/navbar'
 
-import User from './components/user'
-import Users from './components/users'
-import UserCreate from './components/userCreate'
-import UserEdit from './components/userEdit'
-import AssetTransactions from './components/assetTransactions'
-import AssetTransactionsCreate from './components/assetTransactionCreate'
-import Customer from './components/customer'
-import Customers from './components/customers'
-import CustomerCreate from './components/customerCreate'
-import CustomerEdit from './components/customerEdit'
-import Equipment from './components/equipment'
-import Equipments from './components/equipments'
-import EquipmentCreate from './components/equipmentCreate'
-import EquipmentEdit from './components/equipmentEdit'
-import Rentals from './components/rentals'
-import RentalCreate from './components/rentalCreate'
+import User from './components/users/user'
+import Users from './components/users/users'
+import UserCreate from './components/users/userCreate'
+import UserEdit from './components/users/userEdit'
+import AssetTransactions from './components/assetTransactions/assetTransactions'
+import AssetTransactionsCreate from './components/assetTransactions/assetTransactionCreate'
+import Customer from './components/customers/customer'
+import Customers from './components/customers/customers'
+import CustomerCreate from './components/customers/customerCreate'
+import CustomerEdit from './components/customers/customerEdit'
+import Equipment from './components/equipment/equipment'
+import Equipments from './components/equipment/equipments'
+import EquipmentCreate from './components/equipment/equipmentCreate'
+import EquipmentEdit from './components/equipment/equipmentEdit'
+import Rentals from './components/rentals/rentals'
+import RentalCreate from './components/rentals/rentalCreate'
 
 class App extends React.Component {
