+++ react-ui/src/components/customers/customerCreate.js
@@ -2,26 +2,26 @@ import React from 'react'
 import { connect } from 'react-redux'
 import { withRouter } from 'react-router-dom'
 import { Form, Input, Select, Button } from 'semantic-ui-react'
-import ViewHeader from './structure/viewHeader'
-import LinkButton from './structure/linkButton'
-import { updateCustomer } from '../reducers/customerReducer'
+import ViewHeader from '../structure/viewHeader'
+import LinkButton from '../structure/linkButton'
+import { createCustomer } from '../../reducers/customerReducer'
 
 const formStyle = {
   marginTop: 10
 }
 
-class CustomerEdit extends React.Component {
+class CustomerCreate extends React.Component {
   state = {
-    lastName: this.props.initialCustomer.lastName,
-    firstNames: this.props.initialCustomer.firstNames,
-    company: this.props.initialCustomer.company,
-    email: this.props.initialCustomer.email,
-    phone: this.props.initialCustomer.phone,
-    street1: this.props.initialCustomer.billingAddress.street1,
-    street2: this.props.initialCustomer.billingAddress.street2,
-    zip: this.props.initialCustomer.billingAddress.zip,
-    city: this.props.initialCustomer.billingAddress.city,
-    country: this.props.initialCustomer.billingAddress.country,
+    lastName: '',
+    firstNames: '',
+    company: '',
+    email: '',
+    phone: '',
+    street1: '',
+    street2: '',
+    zip: '',
+    city: '',
+    country: '',
     isInvoicable: false
   }
 
@@ -31,25 +31,29 @@ class CustomerEdit extends React.Component {
 
   handleSubmit = async (event) => {
     event.preventDefault()
-    let customer = this.props.initialCustomer
-    customer.lastName = this.state.lastName
-    customer.firstNames = this.state.firstNames
-    customer.company = this.state.company
-    customer.email = this.state.email
-    customer.phone = this.state.phone
-    customer.billingAddress.street1 = this.state.street1
-    customer.billingAddress.street2 = this.state.street2
-    customer.billingAddress.zip = this.state.zip
-    customer.billingAddress.city = this.state.city
-    customer.billingAddress.country = this.state.country
-    await this.props.updateCustomer(customer)
+    const customer = {
+      lastName: this.state.lastName,
+      firstNames: this.state.firstNames,
+      company: this.state.company,
+      email: this.state.email,
+      phone: this.state.phone,
+      billingAddress: {
+        street1: this.state.street1,
+        street2: this.state.street2,
+        zip: this.state.zip,
+        city: this.state.city,
+        country: this.state.country
+      },
+      isInvoicable: this.state.isInvoicable
+    }
+    this.props.createCustomer(customer)
     this.props.history.push('/customers')
   }
 
   render() {
     return (
       <div>
-        <ViewHeader text={`Edit customer ${this.props.initialCustomer.displayName}`} />
+        <ViewHeader text={'Add a customer'} />
         <LinkButton text={'Cancel'} to={'/customers'} type={'default'} />
         <Form style={formStyle} onSubmit={this.handleSubmit}>
           <Form.Field required control={Input} width={6} label='Last name' name='lastName'
@@ -79,19 +83,12 @@ class CustomerEdit extends React.Component {
             </Button>
           </Form.Field>
         </Form>
-      </div>
+      </div >
     )
   }
 }
 
-const mapStateToProps = (store, ownProps) => {
-  const initialCustomer = store.customers.find(c => c._id === ownProps.match.params.id)
-  return {
-    initialCustomer
-  }
-}
-
 export default withRouter(connect(
-  mapStateToProps,
-  { updateCustomer }
-)(CustomerEdit))
+  null,
+  { createCustomer }
+)(CustomerCreate))
